[GEOS-9938] Split the Features template module in independent parts. Moving WFS specific bits to the WFS specific module
