+++ src/community/features-templating/features-templating-core/pom.xml
@@ -47,11 +47,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.geoserver</groupId>
-            <artifactId>gs-wfs</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.github.jsonld-java</groupId>
             <artifactId>jsonld-java</artifactId>
             <version>0.12.5</version>
