+++ src/community/features-templating/features-templating-wfs/src/main/resources/applicationContext.xml
@@ -4,9 +4,14 @@
 
 <beans>
 
-    <bean id="jsonCallback" class="org.geoserver.featurestemplating.request.JsonTemplateCallback">
+    <bean id="jsonCallback" class="org.geoserver.featurestemplating.wfs.JsonTemplateCallback">
         <constructor-arg ref="geoServer"/>
         <constructor-arg ref="templateConfiguration"/>
     </bean>
 
+    <bean id="JSONLDGetFeatureResponse" class="org.geoserver.featurestemplating.wfs.JSONLDGetFeatureResponse">
+        <constructor-arg ref="geoServer"/>
+        <constructor-arg ref="templateConfiguration"/>
+    </bean>
+    
 </beans>
